Remove selected pages from a file

If the toolkit Pdftk is available in the system, it will be called to remove the given pages from the seleted PDF files.

See the reference for detailed usage of pdftk.

remove_pages(rmpages, input_filepath = NULL, output_directory = NULL,
  output_filename = "trimmed_pdf", page_length = NULL)

Arguments

rmpages
a vector of page numbers to be removed

input_filepath
the path of the input PDF file. The default is set to NULL. IF NULL, it prompt the user to select the folder interactively.

output_directory
the path of the output directory

output_filename
the name of the output file.

page_length
page count of the original file

Value

this function returns a PDF document with the remaining pages

References

https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/

Examples

# NOT RUN {
# This command promts the user to select the file interactively.
# Remove page 2 and 3 from the selected file.
remove_pages(rmpages = c(3,6))
# }

dir <- tempdir()
require(lattice)
for(i in 1:3) {
pdf(file.path(dir, paste("plot", i, ".pdf", sep = "")))
print(xyplot(iris[,1] ~ iris[,i], data = iris))
dev.off()
}
staple_pdf(input_directory = dir, output_directory = dir, output_filename = "Full_pdf")

#> Warning: running command 'pdftk "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/Full_pdf.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 1.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 10.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 11.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 12.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 13.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 14.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 15.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 16.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 17.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 18.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 19.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 2.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 20.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 21.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 22.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 23.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 24.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 25.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 26.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 27.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 28.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 29.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 3.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 30.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 31.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 32.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 33.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 34.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 35.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 36.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 37.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 38.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 39.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 4.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 40.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 41.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 42.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 43.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 44.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 45.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 46.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 47.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 48.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 49.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 5.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 50.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/file 51.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0001.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0002.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0003.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0004.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0005.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0006.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0007.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0008.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0009.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0010.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0011.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0012.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0013.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0014.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0015.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0016.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0017.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0018.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0019.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0020.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0021.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0022.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0023.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0024.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0025.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0026.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0027.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0028.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0029.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0030.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0031.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0032.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0033.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0034.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0035.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0036.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0037.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0038.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0039.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0040.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0041.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0042.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0043.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0044.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0045.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0046.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0047.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0048.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0049.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0050.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0051.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0052.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0053.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0054.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0055.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0056.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0057.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0058.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0059.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0060.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0061.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0062.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0063.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0064.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0065.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0066.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0067.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0068.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0069.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0070.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0071.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0072.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0073.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0074.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0075.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0076.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0077.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0078.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0079.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0080.pdf" "C:\Users\Priyanga\AppData\Local\Temp\RtmpWoKQtr/page_0081.pdf" "C:\Users\P

remove_pages(rmpages = c(1), input_filepath = file.path(dir, paste("Full_pdf.pdf",  sep = "")),
 output_directory = dir, page_length = 3)