Remove selected pages from a file

If the toolkit Pdftk is available in the system, it will be called to remove the given pages from the seleted PDF files.

See the reference for detailed usage of pdftk.

remove_pages(rmpages, input_filepath = NULL, output_directory = NULL,
  output_filename = "trimmed_pdf", page_length = NULL)

Arguments

rmpages
a vector of page numbers to be removed

input_filepath
the path of the input PDF file. The default is set to NULL. IF NULL, it prompt the user to select the folder interactively.

output_directory
the path of the output directory

output_filename
the name of the output file.

page_length
page count of the original file

Value

this function returns a PDF document with the remaining pages

References

https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/

Examples

# NOT RUN {
# This command promts the user to select the file interactively.
# Remove page 2 and 3 from the selected file.
remove_pages(rmpages = c(3,6))
# }